Implement `Deas.init`, require routes file and build sinatra app
This implements `Deas.init`. This method requires in the
configured routes file (defaults to `config/routes`) and then
will build a Sinatra modular app, using the `Server` singleton.
The intent is that the app should be defined using the `Server`
singleton and then Deas will handle building the Sinatra app.
`Deas.init` is intended to be called from a `config.ru` before
running `Deas.app`.