Add a `logger` option to `Deas::Server`
This adds a `logger` option to the `Deas::Server` and will set
it on the Sinatra app. This is not a build-in Sinatra option,
so it will have no effect with Sinatra. This is setting a custom
option on the Sinatra app that will store the settings and be
able to hand them off to view handlers when route handle is
implemented.

I started implementing the view handler logic and saw that the
logger is used throughout (based on Sanford's implementation).
Thus, I wanted to add this to go ahead and have it available.