Add ability to define routes using `Deas::Server`
This adds the ability to define routes using `Deas::Server`.
These will be used to build Sinatra routes that run a Deas
view handler. This will allow users to define what HTTP requests
the app will serve. This only contains the ability to define
these routes, but they will not run at this point. The logic
for defining and running view handlers will be added in a future
commit.