Symbolize template names with `render`
The helper method `render` should automatically symbolize
template names. This is a Sinatra feature where if a symbol
is passed to an engine method (like `erb`) then it assumes there
is a template. If it's a string, then Sinatra assumes it is an
inline template. The problem is templates are typically something
like: `"users/index"` and having to manually convert this to a
symbol is annoying. Deas doesn't support inline templates, so we
can always change this to a symbol.