Add nested layouts support to Deas
This adds nested layout support to Deas. Through Sinatra, Deas
supported a single layout, but now it will use Sinatra's
interface for nested layouts. This can be done by passing a
`:layout` or `:layouts` option to the render method. These
are then rendered left-to-right, so the left-most layout is
the outermost layout. Additionally, Sinatra's block style for
nested layouts can be used with the `render` helper. This will
allow view handlers to render their templates in a variety of
ways and also support rendering them within many layouts.