Add a `render` helper to the `RenderScope`
This adds a `render` helper to the `RenderScope`. Similarly to
the `partial` helper, this is used to render templates. The
difference is that this exposes the full flexibility of Sinatra's
template rendering. The primary purpose of this helper is to be
used to render templates in locations outside of the configured
views path or to render non-partial templates.