Add additional helpers for redirecting and using the session
This adds view handler methods for redirecting and using the
session. This allows users to easily redirect requests using
either `redirect` or `redirect_to`. The `redirect_to` method
uses Sinatra's `to` helper method, the `redirect` method doesn't.
In addition to redirecting, if sessions are enabled, the
session can be accessed using `session`. This works the same way
it does in Sinatra. This exposes more of Sinatra's features to
the user through Deas view handler pattern.

Closes #12