Add error handling by configuring error procs on `Deas::Server`
This adds error handling by allowing setting error procs on
the `Deas::Server`. These procs are triggered anytime Sinatra
would trigger it's `error` or `not_found` handlers. They are
passed the exception and evaluated in Sinatra's context, which
allows doing anything possible with Sinatra. This allows apps to
setup error handling easily with the all the features of
Sinatra.

Closes #21