Change Deas logging to use a middleware, not before/after filters
This changes Deas logging to use a middleware instead of using
before and after filters. This was done because before and after
filters were overwriting each other, which made it impossible
to have multiple routes with the same path (but different verbs).
Logging still needs to behave as it did in the filters, ensuring
it runs even if the request halts or raises an exception. Thus,
a middleware.

This implements 2 middleware, one for verbose logging and one
for summary logging. The summary logging should be more
performant and is intended for production environments.

Closes #27 and #28