log errors on both summary and verbose logs
This seems more useful than only logging errors on verbose in general.

But with Sinatra's current implementation, `env['sinatra.error']`
isn't even set when `show_exceptions` is set.  This side-effect of
this is that, b/c in typical development setups, `show_exceptions`
and `verbose_logging` are both set to `true`, this has no effect.

Now, this is OK when you are showing exceptions, but if you don't
also log exceptions in summary logging, then the only configuration
where this would be used is the case where you have `show_exceptions`
set to `false` and `verbose_logging` set to `true`.  This seems like
an edge-configuration to me.

This opens up error logging to the (I hope more common) production
configuration of `show_exceptions=false` and `verbose_logging=false`.