Add `redirect` route helper for building routes that redirect
This adds a `redirect` helper for `Deas::Server`'s that allows
allows easily redirecting routes without having to define a
view handler. This is essentially a macro that builds the view
handler for the route and redirects in it's `run!` method.

This also removes the `redirect_to` helper. After looking in
Sinatra's code, using `redirect to('/path')` and
`redirect '/path'` are equivalent. Because of this, the
`redirect_to` helper doesn't add any extra functionality. Thus,
this removes it to have a smaller interface to maintain.