Change Deas usage of Sinatra `show_exceptions` setting
Sinatra's `show_exceptions` setting (and it's `raise_errors`
setting) are not compatible with Deas. They interfere with Deas'
logging and prevent it's error handler's from running. This is
odd behavior and is confusing when encountered in development.
To fix this, `show_exceptions` and `raise_errors` are always
disabled. Deas will now add a middleware that will show
exceptions similar to how Sinatra's `show_exceptions` worked if
the `show_exceptions` Deas option is configured for a server.