template engine: look in `:views` option first, fallback to sinatra views
This alters the algorithm for looking up a named templates rendering
engine.  Before, if the template had a custom `:views` option, it
would be ignored in this algorithm.

This alters to first look in the templates `:views` option and fallback
to the main sinatra `:views` setting if no custom option is specified.
This enables plugins to define their own template files and have their
engine properly be detected.

No major behavior changes to the engine rendering scheme - this is
just a patch to the previous effort.