auto-detect and set the content type based on the template extname
This updates the sinatra runner's `render` method to detect and set
the content type based on the template name `extname`.  This provides
a way to specify content type based on template naming conventions.

If the template name has no `extname`, `html` is assumed.  Also,
any other metho of setting the content type will take precedence
over this so that custom content types not matching the convention
can still be used.