specify url names on routes
This adds the (optional) ability to specify a url name when devining
routes.  This will configure a `Url` object for that name with the
given route's path and stored on the server's configuration.

Urls are only created if a url name is given and the path for the
route is given as a `String`.  This url object will eventually be
used to generate route strings given the name and some optional
parameters.

Also, this contains a bunch of "cleanups" to the options on the server
configuration as well.  This is just to better organize them (there
are a lot) and help with readability.