ensure a named url exists when configuring a redirect to it
Because we need to lookup the Url at redirect definition time, we
need to ensure the the Url has been defined before we can setup a
redirect to it.  We raise a helpful exception when the Url doesn't
exist to help the developer understand what is going on.

Now, this isn't ideal.  It forces url names to be specified as a
Symbol and it forces you to define your urls before defining the
redirects to them.  There isn't really anything to do about this -
just wanted to point it out in case a better implementation comes
up later.