Minor test tweaks for Rubinius 2.0 compatibility
This tweaks the test suite for compatibility with Rubinius. This
required only one change to the test suite. We were relying on
`Array#to_s` which behaves different between MRI and Rubinius.
This modifies the test to not use the `to_s` method but still
keeps the integrity of the test intact. This also manually
includes "haml" to avoid a warning from tilt.