render partials from the view handlers
This adds the ability to render partials from the view handlers.
This follows the pattern of the runners implementing the specific
behavior and the view handler demetering to its runner.

This overarching goal here is to render partial html content to
expose as view handler attrs and include in things like JSON data
responses.