view handler: expose the router as a protected helper method
This updates the view handler to have the router exposed as a
protected helper method.  This follows the pattern of other helpers
like the logger method.  The goal here is to allow plugins to do
things with the router.  To do this the router needs to be provided
in a generic, programmable way.

To accomplish this, we add the router to the sinatra call and pipe
it to the view handler via the runners.