remove `partial` method from view handlers and runners
This isn't a universal capability across template engines so it
shouldn't be available at the view handler level.  Template engines
should be the ones to provide this function if they choose to.

This was really just a macro for a special `render` call anyway so
we aren't really losing any capabilities.

In a future PR, the `partial` method and `Partial` template subclass
will need to be removed from "general" deas and into an erb template
engine gem (once the API for template engines is ready).  Again,
this isn't a global template engine capability so it shouldn't be
exposed as one.

This is all prep for building a template rendering API and not
relying on Tilt/Sinatra.