pass standard args in to the sinatra runner
This removes auto setting standard args from the sinatra call in the
sinatra runner.  This 1) removes logic from the sinatra runner as
we are moving away from it and 2) allows for setting arguments (like
logger and router) without needing the sinatra call (again, moving
away from relying on sinatra).

This is all setup and work towards phasing out Sinatra and replacing
with native equivalent logic.