break out the null logger into its own file; update requires
This moves the null logger definition to its own file and updates
references to specifically require this in.  This is more about
cleanup and consistency - no behavior changes.

This is also setup for future logging changes.  There will most likely
be a logger that encapsulates the verbose/summary logging logic
currently living in the logging middleware.  Who knows, the middleware
may dissappear in the future.

Anyway, this is cleanup and prep for those possible future changes.