sinatra runner: render with template source if there is an engine
This updates the sinatra runner to prefer to render using a matching
engine on the template source if one has been configured.  If not
it falls back to rendering with Sinatra as before.

This is a temporary patch to allow transitioning off sinatra and
on to engines.  This is part of moving away from Sinatra for rendering.

Note too that I removed the alt layout render system tests b/c we
are no longer supporting calling `render` with a block in the view
handlers.