have `compile` render api expect the template name
This is so you can provide meaningful backtraces and file information
even when rendering content that has already been rendered by another
engine.

This came up when attempting to implement `compile` for deas-erubis.
It is nice to give a file name so any exceptions get nice backtraces.

Note: this also removes locals from the expected API.  Locals are
only needed at the top-level engine as all local handling should
be handled with that engine (ie Erb or other).  Compiling with
downstream engines should be about converting a more convenient
markup (ie writing markdown with erb) to the needed markup.