template source: render honoring any view handler layouts
This updates the template source to look for any layouts on the
view handler and build nested blocks that render each template
file.  This allows for templates to yield in each layout template
and render layouts.

This also updates the template engine render api to expect a content
block.  This is needed so the template engines can pass the layout
render blocks to their render logic and implement yielding in their
templates.