render API: remove `capture_partial` method
This updates the render API to support sending a content block to
source/engine partial calls.  Doing this removes the need for a
dedicated capture partial method as it is just a partial call with
a content block.

This also cleans up the API to formally allow passing content blocks
to both the `render` and `partial` methods at the source level.  The
runner doesn't expose this to the view handler as it doesn't make
sense.  However, engines can use this in their template helpers.

This is all part of standardizing a rendering API that allows for
creating specific template engines that support rich rendering.
Deas::Erubis will specifically take advantage of these APIs to
allow rich partial rendering.