Remove render logic from the sinatra runner; switch to always using template source
This updates to remove rendering logic from the sinatra runner.  This logic now
moves to the deas runner (with the exception of auto content type handling).
This deprecates the sinatra render API and switches to always using the template
scope for rendering, including:

* remove all template modeling and scope handling
* remove muli json and yajl dependencies - instead just inspect the hash and compare
* remove haml dependency - not needed since no longer rendering in tests

This is part of removing Sinatra as a dependency of Deas.