cache template name extensions at the template source
This is a small render optimization.  The goal is to limit the
somewhat expensive logic of globing files on disk that match the
template name, rejecting out irrelevent ones and choosing the first
matching one just in order to lookup a files extension.

With this implementation, the expensive logic is only run once per
template (on the first lookup).  This will improve every render
call as this lookup is done both at engine-selection time but also
when querying to know if there is an engine available for the template.

Note: this also adds a test to make sure `engine_for?` returns false
on missing template names.  This was missing in the original tests
and is all about better coverage.