rework template source api for querying about configured engines
This reworks the `engine_for?` method to just take an extension
as an argument.  It then adds an `engine_for_template?` method that
works like the prev engine for method did.

The goal here is to be able to query whether an engine has been
configured when you already know the extension while keeping the
old ability to query when you don't know the extension but know the
template.