hotfix: switch back to more verbose runner ivar in the view handler mixin
This was accidentally changed in the previous PR #165.  This ivar
needs to be prefixed with `deas_` so as not to "pollute" the ivar
names on view handlers.  `@runner` is too generic and may need to
be available for the users.  This reverts the ivar to how it was
named before.

/cc @jcredding