template source: rename `default_template_source` engine option
This is a simple rename.  This name better describes the intent and
purpose of this option.  It is the source that will be used to
render if no specific source is specified.

This is useful if configuring 3rd-party template sources.  You may
want to configure their sources to use your app's template source
by default so that any generic renders done within 3rd-party templates
behave the same as if they were done within your app.  A case for
this would be if a 3rd-party template needs to render app-specific
partials.