have the test runner render templates and discard
The goal here is better test coverage and a better chance at
catching errors in templates.  The test runner will now render the
actual templates, just as the deas runner would, and then discard
any rendered output.  This does mean that a valid template source
needs to be configured and setup for the test suite, just as it is
for the server.

To accomplish this change, I moved the render methods to the base
runner.  The deas runner no longer implements them and the test
runner supers to them before building its render args as before.

Note: I also chose to simplify the render args handling and just
use a single render args class that all render methods build.
This also includes some tweaks and cleanups to the test suite
necessary to render the templates in the tests.