router: api to route same meth/path to diff handlers based on request
This updates the API for defining routes to allow routing the same
method/path combination to different handlers based on the type of
request.  The goal here is to support rendering both mobile and
desktop versions of the same route.

A side-effect of this api change is you are no longer *forced* to
specify a handler when defining your routes.  In this case Deas will
error nicely.  Another side effect is handling the "default" case.
This involves optionally setting a default request type and handling
the case when no default handler is specified.

Note: this cleaned up the router/route/proxy handling too.  There
was a lot of coupling and who-handled-what didn't make a lot of sense.
These objects have been updated to better reflect their intended
responsibilities.  The associated tests are cleaned up and modernized
a bit too.