view handlers: have `layout` DSL take procs that are instance eval'd
This is so you can use handler attributes to determine your layout
template paths.  This allows for dynamic layout templates based on
the nature of the request.

Note: this also changes the API of the layout method to no longer
take multiple paths.  It now expects either a single path or block.
This also removes the `layouts` alias.  Now, layouts just returns
the layouts that have been specified.  This is to simplify the layout
API now that it is taking procs.  Plus, the multiple layout template
path feature hasn't been used much in practice.

Note also: the view handler instance now has a `layouts` method that
returns the layout templates paths.  It handles instance eval'ing
the class layout procs against the handler instance.